tools/social_publisher.py
#!/usr/bin/env python3
"""
China Social Media Automation Toolkit
今日头条自动发布工具
"""

import os
import time
import json
import logging
from datetime import datetime
from pathlib import Path

import yaml
import requests
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.common.exceptions import TimeoutException

# 配置日志
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

# 配置文件
CONFIG_FILE = Path(__file__).parent / "config.yaml"


class ToutiaoPublisher:
    """头条号发布器"""
    
    def __init__(self, phone: str = None, password: str = None):
        self.phone = phone or os.getenv("TOUTIAO_PHONE")
        self.password = password or os.getenv("TOUTIAO_PASSWORD")
        self.base_url = "https://www.toutiao.com"
        self.driver = None
        
    def login(self) -> bool:
        """登录头条号"""
        if not self.phone or not self.password:
            logger.error("请配置手机号和密码")
            return False
            
        logger.info(f"开始登录头条号: {self.phone}")
        # 这里添加实际的登录逻辑
        return True
    
    def publish(self, content: str, image_path: str = None) -> dict:
        """发布微头条"""
        logger.info(f"发布内容: {content[:50]}...")
        
        # 1. 生成或上传图片
        image_url = None
        if image_path:
            image_url = self._upload_image(image_path)
        elif content:
            # 使用AI生成配图
            image_url = self._generate_image(content)
            
        # 2. 发布内容
        # 这里添加实际的发布逻辑
        
        return {
            "success": True,
            "content": content,
            "image": image_url,
            "time": datetime.now().isoformat()
        }
    
    def _upload_image(self, image_path: str) -> str:
        """上传图片"""
        # 实际实现需要Selenium或API
        logger.info(f"上传图片: {image_path}")
        return image_path
    
    def _generate_image(self, content: str) -> str:
        """使用AI生成配图"""
        # 这里可以集成OpenAI DALL-E或其他图像生成API
        logger.info(f"为内容生成配图: {content[:30]}...")
        return None
    
    def get_stats(self) -> dict:
        """获取账号统计数据"""
        return {
            "fans": 150,
            "views": 0,
            "earnings": 1.37
        }


class XiaohongshuPublisher:
    """小红书发布器"""
    
    def __init__(self, phone: str = None, password: str = None):
        self.phone = phone or os.getenv("XIAOHONGSHU_PHONE")
        self.password = password or os.getenv("XIAOHONGSHU_PASSWORD")
        
    def publish(self, title: str, content: str, images: list = None) -> dict:
        """发布小红书"""
        logger.info(f"发布笔记: {title}")
        
        # 生成标题、标签
        tags = self._generate_tags(content)
        
        return {
            "success": True,
            "title": title,
            "content": content,
            "tags": tags,
            "images": images
        }
    
    def _generate_tags(self, content: str) -> list:
        """生成热门标签"""
        # 简单实现，实际可以用AI
        keywords = ["#", "#生活", "#分享"]
        return keywords


class SocialMediaManager:
    """社交媒体管理器"""
    
    def __init__(self, config_path: str = None):
        self.config = self._load_config(config_path)
        self.publishers = {}
        self._init_publishers()
        
    def _load_config(self, config_path: str = None) -> dict:
        """加载配置"""
        config_file = Path(config_path) if config_path else CONFIG_FILE
        
        if config_file.exists():
            with open(config_file) as f:
                return yaml.safe_load(f) or {}
        return {}
    
    def _init_publishers(self):
        """初始化发布器"""
        # 头条号
        toutiao_config = self.config.get("toutiao", {})
        self.publishers["toutiao"] = ToutiaoPublisher(
            phone=toutiao_config.get("phone"),
            password=toutiao_config.get("password")
        )
        
        # 小红书
        xhs_config = self.config.get("xiaohongshu", {})
        self.publishers["xiaohongshu"] = XiaohongshuPublisher(
            phone=xhs_config.get("phone"),
            password=xhs_config.get("password")
        )
    
    def publish(self, platform: str, **kwargs) -> dict:
        """发布到指定平台"""
        publisher = self.publishers.get(platform)
        if not publisher:
            return {"success": False, "error": f"不支持的平台: {platform}"}
        
        return publisher.publish(**kwargs)
    
    def get_all_stats(self) -> dict:
        """获取所有平台统计"""
        stats = {}
        for platform, publisher in self.publishers.items():
            if hasattr(publisher, 'get_stats'):
                stats[platform] = publisher.get_stats()
        return stats


def main():
    """主函数"""
    import argparse
    
    parser = argparse.ArgumentParser(description="中国社交媒体自动化工具")
    parser.add_argument("--platform", choices=["toutiao", "xiaohongshu"], required=True)
    parser.add_argument("--content", help="发布内容")
    parser.add_argument("--title", help="标题(小红书)")
    parser.add_argument("--image", help="配图路径")
    parser.add_argument("--stats", action="store_true", help="查看统计")
    
    args = parser.parse_args()
    
    manager = SocialMediaManager()
    
    if args.stats:
        stats = manager.get_all_stats()
        print(json.dumps(stats, indent=2, ensure_ascii=False))
    else:
        result = manager.publish(args.platform, content=args.content, title=args.title)
        print(json.dumps(result, indent=2, ensure_ascii=False))


if __name__ == "__main__":
    main()