requirements.txt
# Web 界面
streamlit>=1.28.0

# 浏览器自动化
selenium>=4.15.0
playwright>=1.40.0

# AI
openai>=1.3.0
anthropic>=0.18.0

# 数据处理
pandas>=2.1.0
python-dotenv>=1.0.0

# 定时任务
apscheduler>=6.10.0

# HTTP 请求
requests>=2.31.0
httpx>=0.25.0

# JSON/YAML
pyyaml>=6.0.0

# 日志
loguru>=0.7.0

# 加密货币（可选）
ccxt>=3.2.0
hyperliquid-sdk>=1.0.0